moved entities folder out of beans
