+++ Spring/src/main/java/com/revature/Group4P2/beans/repositories/CatalogDetailsRepo.java
@@ -1,6 +1,6 @@
 package com.revature.Group4P2.beans.repositories;
 
-import com.revature.Group4P2.beans.entities.CatalogDetails;
+import com.revature.Group4P2.entities.CatalogDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CatalogDetailsRepo extends JpaRepository<CatalogDetails, Integer> {
