+++ Spring/src/main/java/com/revature/Group4P2/beans/repositories/PurchasesRepo.java
@@ -1,6 +1,6 @@
 package com.revature.Group4P2.beans.repositories;
 
-import com.revature.Group4P2.beans.entities.Purchases;
+import com.revature.Group4P2.entities.Purchases;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PurchasesRepo extends JpaRepository<Purchases, Integer> {
