+++ Spring/src/main/java/com/revature/Group4P2/entities/Catalog.java
@@ -1,14 +1,11 @@
-package com.revature.Group4P2.beans.entities;
+package com.revature.Group4P2.entities;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 import javax.persistence.*;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Objects;
 
-@Entity
-@Table(name = "catalog")
+@Entity(name = "catalog")
 public class Catalog {
 
     @Column(name = "item_id")
