+++ Spring/src/main/java/com/revature/Group4P2/entities/Purchases.java
@@ -1,12 +1,11 @@
-package com.revature.Group4P2.beans.entities;
+package com.revature.Group4P2.entities;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 import javax.persistence.*;
 import java.util.Objects;
 
 
-@Entity
-@Table(name = "purchases")
+@Entity(name = "purchases")
 public class Purchases {
 
     @Id
