+++ Spring/src/main/java/com/revature/Group4P2/entities/Users.java
@@ -1,14 +1,12 @@
-package com.revature.Group4P2.beans.entities;
+package com.revature.Group4P2.entities;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import javax.persistence.*;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
 
-@Entity
-@Table(name = "shopping_user")
+@Entity(name = "shopping_user")
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
